Renamed global data provider to PostAnalytics
