+++ apps/posts/src/providers/PostAnalyticsContext.tsx
@@ -4,7 +4,7 @@ import {STATS_RANGES} from '@src/utils/constants';
 import {Setting, useBrowseSettings} from '@tryghost/admin-x-framework/api/settings';
 import {StatsConfig} from '@tryghost/admin-x-framework';
 
-type GlobalDataContextType = {
+type PostAnalyticsContextType = {
     data: Config | undefined;
     statsConfig: StatsConfig | undefined;
     isLoading: boolean;
@@ -15,17 +15,17 @@ type GlobalDataContextType = {
     settings: Setting[];
 }
 
-const GlobalDataContext = createContext<GlobalDataContextType | undefined>(undefined);
+const PostAnalyticsContext = createContext<PostAnalyticsContextType | undefined>(undefined);
 
 export const useGlobalData = () => {
-    const context = useContext(GlobalDataContext);
+    const context = useContext(PostAnalyticsContext);
     if (!context) {
-        throw new Error('useGlobalData must be used within a GlobalDataProvider');
+        throw new Error('useGlobalData must be used within a PostAnalyticsProvider');
     }
     return context;
 };
 
-const GlobalDataProvider = ({children}: { children: ReactNode }) => {
+const PostAnalyticsProvider = ({children}: { children: ReactNode }) => {
     const config = useBrowseConfig();
     const [range, setRange] = useState(STATS_RANGES.LAST_30_DAYS.value);
     const settings = useBrowseSettings();
@@ -41,7 +41,7 @@ const GlobalDataProvider = ({children}: { children: ReactNode }) => {
         throw error;
     }
 
-    return <GlobalDataContext.Provider value={{
+    return <PostAnalyticsContext.Provider value={{
         data: config.data as Config | undefined,
         statsConfig: config.data?.config?.stats as StatsConfig | undefined,
         isLoading,
@@ -52,7 +52,7 @@ const GlobalDataProvider = ({children}: { children: ReactNode }) => {
         settings: settings.data?.settings || []
     }}>
         {children}
-    </GlobalDataContext.Provider>;
+    </PostAnalyticsContext.Provider>;
 };
 
-export default GlobalDataProvider;
+export default PostAnalyticsProvider;
