+++ apps/posts/src/routes.tsx
@@ -2,6 +2,7 @@ import Growth from './views/PostAnalytics/Growth/Growth';
 import Newsletter from './views/PostAnalytics/Newsletter/Newsletter';
 import Overview from './views/PostAnalytics/Overview/Overview';
 import PostAnalytics from './views/PostAnalytics/PostAnalytics';
+import PostAnalyticsProvider from './providers/PostAnalyticsContext';
 import Web from './views/PostAnalytics/Web/Web';
 import {ErrorPage} from '@tryghost/shade';
 import {RouteObject} from '@tryghost/admin-x-framework';
@@ -20,8 +21,14 @@ export const routes: RouteObject[] = [
         errorElement: <ErrorPage onBackToDashboard={() => {}} />, // @TODO: add back to dashboard click handle
         children: [
             {
+
+                // Post Analytics
                 path: 'analytics/beta/:postId',
-                element: <PostAnalytics />,
+                element: (
+                    <PostAnalyticsProvider>
+                        <PostAnalytics />
+                    </PostAnalyticsProvider>
+                ),
                 children: [
                     {
                         path: '',
@@ -41,6 +48,8 @@ export const routes: RouteObject[] = [
                     }
                 ]
             },
+
+            // Error handling
             {
                 path: '*',
                 element: <ErrorPage onBackToDashboard={() => {}} /> // @TODO: add back to dashboard click handle
