+++ apps/posts/src/views/PostAnalytics/PostAnalytics.tsx
@@ -2,7 +2,7 @@ import PostAnalyticsLayout from './components/layout/PostAnalyticsLayout';
 import {Outlet} from '@tryghost/admin-x-framework';
 import {STATS_RANGES} from '@src/utils/constants';
 import {useEffect} from 'react';
-import {useGlobalData} from '@src/providers/GlobalDataProvider';
+import {useGlobalData} from '@src/providers/PostAnalyticsContext';
 
 const PostAnalytics: React.FC = () => {
     const {setRange} = useGlobalData();
